Move get_module back to parser
